Pipes terminado en teoria
No se testeo todavia

+++ TPE/Userland/SampleCodeModule/commands.c
@@ -27,7 +27,7 @@ void clearScreen(){
 
 
 void time(){
-	uint64_t * currTime = sys_time();
+	int * currTime = sys_time();
 	// uint64_t * currTime = os_time();		returns an array with the current hour, minutes, and seconds
 
 	char time[9];					//pass the array to a string to print
