+++ TPE/Userland/SampleCodeModule/snake.c
@@ -433,7 +433,7 @@ void showScore(){
 
 int getseconds(){
   //uint64_t *currTime = os_time();
-  uint64_t *currTime = sys_time();
+  int *currTime = sys_time();
 
   // printf("%d%d , %d%d, %d%d",currTime[5],currTime[4],currTime[3],currTime[2],currTime[1],currTime[0]);
